ATSP {TSP}	R Documentation

Class ATSP – Asymmetric traveling salesperson problem

Description

Constructor to create an instance of the asymmetric traveling salesperson problem (ATSP) and some auxiliary methods.

Usage

## constructor
ATSP(x, labels = NULL)

## coercion
as.ATSP(x)

## methods
## S3 method for class 'ATSP'
n_of_cities(x)
## S3 method for class 'ATSP'
image(x, order, col = gray.colors(64), ...)
## S3 method for class 'ATSP'
labels(object, ...)
## S3 method for class 'ATSP'
print(x, ...)

Arguments

x, object
an object (a square matrix) to be converted into an ATSP or, for the methods, an object of class ATSP.

labels
optional city labels. If not given, labels are taken from x.

col
color scheme for image.

order
order of cities as an integer vector or an object of class TOUR.

...
further arguments are passed on.

Details

Objects of class ATSP are internally represented by a matrix (use as.matrix() to get just the matrix).

ATSPs can be transformed into (larger) symmetric TSPs using reformulate_ATSP_as_TSP().

Value

ATSP() returns x as an object of class ATSP.

n_of_cities() returns the number of cities in x.

labels() returns a vector with the names of the cities in x.

Author(s)

Michael Hahsler

See Also

TOUR, insert_dummy, tour_length, reformulate_ATSP_as_TSP, solve_TSP.

Examples

data <- matrix(runif(10^2), ncol = 10, dimnames = list(1:10, 1:10))

atsp <- ATSP(data)
atsp

## use some methods
n_of_cities(atsp)
labels(atsp)

## calculate a tour
tour <- solve_TSP(atsp, method = "nn")
tour
  
tour_length(tour)

image(atsp, tour)

[Package TSP version 1.1-6 Index]