TSP-package {TSP}	R Documentation

TSP: Infrastructure for the Traveling Salesperson Problem

Description

Basic infrastructure and some algorithms for the traveling salesperson problem (also traveling salesman problem; TSP). The package provides some simple algorithms and an interface to the Concorde TSP solver and its implementation of the Chained-Lin-Kernighan heuristic. The code for Concorde itself is not included in the package and has to be obtained separately. Hahsler and Hornik (2007) doi: 10.18637/jss.v023.i02.

Key functions

solve_TSP()

Author(s)

Maintainer: Michael Hahsler mhahsler@lyle.smu.edu (ORCID) [copyright holder]

Authors:

Kurt Hornik (ORCID) [copyright holder]

See Also

Useful links:

https://github.com/mhahsler/TSP

Report bugs at https://github.com/mhahsler/TSP/issues

[Package TSP version 1.2.6 Index]