cut_tour {TSP}	R Documentation

Cut a tour to form a path

Description

Cuts a tour at a specified city to form a path.

Usage

cut_tour(x, cut, exclude_cut = TRUE)

Arguments

x
an object of class TOUR.

cut
the index or label of the city to cut the tour.

exclude_cut
exclude the city where we cut? If FALSE, the city at the cut is included in the path as the first city.

Author(s)

Michael Hahsler

See Also

TOUR.

Examples

data("USCA50")

tsp <- insert_dummy(USCA50, label = "cut")
tour <- solve_TSP(tsp)

## cut tour into path at the dummy city
path <- cut_tour(tour, "cut")

labels(path)

[Package TSP version 1.1-6 Index]